move httpapi_schema
